requirements.txt
azure-cosmos
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
propcache==0.4.1
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0